Vim: better time/date insertion
Now the <F1> key inserts a newline with a date if the cursor is not
currently on the first column. This could perhaps further be improved by
detecting whether there are any characters already inserted into the
current line, and only if there are no characters, to insert a newline
before inserting the time/date.