xorg: exelion: dual portrait mode
These few lines change the xorg.conf file so that the two monitors
display a resolution of 2100x1680 (portrait mode) instead of 3360x1050
(landscape mode, aka, normal mode).

The only drawback to this at this point in time is that xmonad is unable
to draw window borders. This is apparently an issue with the nvidia
driver, and not xmonad.