shellscript: alarm: new formula for changing volume
The new formula makes it so that the volume rises much quicker in the first few seconds, and slows down much more later on. The effect is that we don't have to wait 10 minutes for the sound to be recognizable.
also: wakeup time is set at 5:50 am instead of 4:40am.