autocall: complete rewrite with tons of options
This is basically a feature-complete version of autocall, implemented
with zsh. The only zsh-specific thing in autocall, aside from
syntactical differences from other shells, is the builtin "read"
function (not compatible with bash's builtin "read").

There are so many neat options in autocall that it is basically 95%
done for the foreseeable future (the only large domain missing is the
lack of tracking remote files, or perhaps the outputs of other commands
themselves).