vim: only set local buffer's options with autocommand
With the old behavior, we used *set* when opening up files with, e.g.,
*.c extension. This meant that afterwards, if we opened up a file that
wasn't *.c, (such as a plain text file), it would inherit all of the
things set with *set*, such as textwidth.

By using *setlocal* instead, we get around this problem by limiting the
effects of the autocommands to the local buffers only.