xmonad: let amixer hotkeys use discrete dB units
Interestingly, on exelion and aether, the Master volume meter's range is
64 dB units, incremented by 1 dB unit at a time (0dB for 100% volume and
-64dB for 0% volume on the exelion). On luxion, the range is actually
only 46.5dB, but with exact 1.5dB increments instead (so 31 units
total).

Alsamixer can change volume by the dB instead of by flat percentages.
This is much better because it provides exact accuracy by the dB.
Interestingly, we can just use a universal 1.5dB increment for all
systems (exelion, aether, and luxion) because alsamixer nicely uses the
correct smallest dB increment depending on the system. So if we call
with 1.5dB+ on exelion, only 1dB actually gets incremented. On the
luxion, 1.5dB+ causes exactly 1.5dB (the smallest available increment)
of the volume to be increased.

This patch also fixes the annoying bug where the hotkeys did not
actually make the volume go up all the way to 100% or 0% (probably
because alsamixer discarded the command of increasing or decreasing by
3% if the amount to be changed was, e.g., 98% (when increasing) or 2%
(when decreasing)).