xmonad: find empty WS in Group: after moving, spawn a terminal
This makes it function in a conceptually similar manner as Vim's own "o"
hotkey, which creates a new line. Here, mod + alt + "o" moves to the
next blank WS in the Group (if there is such a WS; otherwise it stays in
the same WS) and spawns a terminal.

This hotkey (which takes advantage of the new WS Groups idea) is
probably the newest "killer feature" of my Xmonad config.