xorg/xinitrc/xmonad: use xmonad to switch keyboard layouts
We no longer depend on the xorg configuration files to change the
keyboard layout. Instead, we use the userland command "setxkbmap" (which
is maintained by the same xorg developers) to explicitly change layouts.

This has 2 advantages: (1) Because it's in userspace now, we no longer
have to restart the X server to see if a "XkbLayout" option or some
other has taken effect. (2) Because we it's in userspace, we can use
xmonad to leverage it, which means that we can use any keys we want,
reliably (just as reliable as xmonad itself).

Currently, a problem might occur if the layout_switch.sh script is
called too quickly (thus resulting in the xmodmap.sh calls interleaving
with each other). However, this is not a serious problem.