zsh/script: usbmnt: smarter version
This version of usbmnt is much more intelligent and smarter than before,
with a new menu-driven interface (which is sanely automated when there
is just 1 USB device to mount, or unmount). It parses the output of the
"blkid" program for all necessary information.

The new zsh aliases reflect the new options supported by usbmnt.