zsh: CAPSLOCK-proof alias for initkeys
Tampering with the the keyboard during the init process (when
initkeys.sh is called) sometimes results in the CAPSLOCK being turned on
without any way to turn it off. So even if we wanted to call 'initkeys'
again, we can't do so, because of CAPSLOCK. This alias works around that
scenario.