mutt: simplify password handling
The pass.gpg file now is itself a mutt configuration snippet, so we can
just use the "source" command to includ its contents. This method is
drastically simpler as well.

The unusual benefit is that we use the same amount of quoting for the
metacharacters in the password as in the old method.