vim: fix <space> and <backspace> mappings
For a long time, I did not realize that the "10<c-e>10j" keybinding was
not what I wanted, as it moved the cursor more than 10 lines down (<c-e>
is an "impure" way to scroll). The proper way is to use "zz", which
resets the screen to have the cursor in the middle of the screen
(vertically), <em>without</em> changing the current line position.