zsh: fix 'e' alias
The 'e' alias referred to emacsclient_open(), but this function was
defined together with emacslicent_tty_open(), and resided in the file
emacsclient_open. This resulted in a silent error that Zsh did not
report, and required calling emacsclient_open() two times to get it to
work as expected. By moving each function into its own file, we avoid
this error.