zsh: k2: stop lying about TERM
This is an ugly fix (even if it works on NixOS), because it does not
even consider if the current OS is NixOS. Since we don't use NixOS on k2
anyway, remove this.

This should also prevent urxvt from hanging/freezing on its own somewhat
randomly (I thought that this was an intel video driver bug for the past
few months --- oops!).