emacs: evil/haskell-mode: disable annoying behavior from "S" or "cc"
This disables the evil-auto-indent variable whenever haskell-mode is
activated. The (evil-change-whole-line) function, which is mapped by "S"
or "cc", eventually calls (evil-change), which calls (evil-open-above),
which calls (indent-according-to-mode) if evil-auto-indent is defined
(not nil). Setting evil-auto-indent to nil bypasses the call to
(indent-according-to-mode); this way, haskell-mode's
(haskell-mode-suggest-indent-choice) does not get triggered
(haskell-mode forces the user to select a predefined indentation mode
provided by haskell-mode if none is selected, and brings up the
documentation in a new buffer for haskell-mode-hook).