emacs: haskell-mode: simplify indentation code
This removes lots of code while maintaining the same behavior as before.
The "S" and "cc" commands work perfectly fine, too, so this obsoletes
commit 018119cebce23379117505753dce8c8598dd87ba. The key change is is

    (setq indent-line-function #'indent-relative)

which is all it takes to remove the custom (newline-and-indent-relative)
function and the special keybindings in haskell-mode for "RET", "o", and
"O", while also maintaining the proper behavior of "S" and "cc".

No obvious functional change.