emacs: toggle "kj" mapping when UIM mode is active
The "kj" binding uses a special function that captures the "k" character
from INSERT mode; this is fine for English, but when inputting Japanese,
the "k" needs to be considered as part of a Japanese syllabary for
candidate selection by the Anthy input engine. We use the new function
(my-uim-mode) to toggle the "k" mapping off whenever UIM mode is active.