emacs: fix tiny font problem on k1
The solution is to use the `system-name` parameter. To see the value of
this parameter in Emacs, use M-: in Emacs mode to see the value printed
in the echo area. Also, we have to modify the quoted call in
`custom-set-faces` to a backquote, to get this to work without Elisp
errors. See
http://stackoverflow.com/questions/10573219/interesting-error-with-emacs-lisp
.