emacs: insert-stupid-tab
This is my own stupid version of the "Smart Tabs" mode idea from the
Emacs community. "Smart Tabs" is a mode and requires you to write helper
functions for the major mode in use; my stupid version relies on the
human user for aesthetics.

This function inserts a TAB when we are indenting (leading whitespace),
or the right number of SPACE characters when we are aligning something
inside or at the end of a line.