emacs: kakapo-project-settings: fix 'text-mode-hook
In our emacs/init.el file, we have this:

    (add-hook 'prog-mode-hook 'my-kakapo-indents)
    (add-hook 'text-mode-hook 'my-kakapo-indents)

. This was broken when we did "emacs some/file.rb", because doing so
only ran the `prog-mode-hook`, and so, adding `my-kakapo-indents` to
`text-mode-hook` called it, but the "buffer-file-name" was nil because
the buffer name only becomes non-nil if we are in the prog-mode-hook,
not text-mode-hook. This resulted in a stringp error message (for
variable 'b'), because emacs still tried to run the value of b (nil)
against the string-match functions in 'my-kakapo-indents'.