emacs: hiwin bg color: unify gui/terminal behavior
As we use the zenburn theme in terminal emacs (not zenmonk), we need to
adjust the inactive background to be brighter than "gray20"; at the same
time, if we just use "gray20" for both terminal (zenburn) and gui
(zenmonk) emacs, we get inconsistent behavior (in terminal, inactive
color is darker than active color, but in gui, inactive color is
brighter).  Tweaking the values this way makes the inactive buffer
colors brighter in both terminal/gui emacs.