emacs: enhance undoing of kakapo-open
There are many times when we want to undo pressing "o" or "O"
`kakapo-open`, but that involves multiple presses of "u" for undo after
escaping. This makes it so that if we escape to normal mode from
`kakapo-open` without inserting a non-whitespace character, we undo
repeatedly until we find the line that point originated from, before
`kakapo-open`.

Because this is experimental, we leave it outside of kakapo, although it
might make sense to mention this tweak in kakapo itself eventually.