nixos: k0: additional X-related settings
`desktopManager.xterm.enable = false;` is to disable the initial xterm,
which is incorrectly set as the session (which results in opening an
xterm, which, if closed, kills the X session). The old "workaround" was
to type in "xmonad" into this xterm, and then use XMonad for a bit, but
disabling xterm directly is the correct solution.

`xset` is used for disabling mouse acceleration.