nixos: w0 configuration
This is the minimum configuration required to get NixOS running as a guest in
VirtualBox. The only quirk is setting up the normal user 'l'. I am not aware of
a way to set up a group named 'l', so I had to use `groupadd` and `usermod` to
create the 'l' group.

The (virtualized) ethernet connection seems to be working fine without any
issues (without bridged networking from VirtualBox).