urxvt: diable login shell
With the old setting, any shell started within it is treated as a login
shell. E.g., if we do "urxvt" and our user's default shell is set to
"zsh", then "zsh" thinks that it is a login shell and proceeds to
evaluate ~/.zprofile (followed by ~/.zshrc, and then finally ~/.zlogin).