nixos: enable NTP
Strangely, ntp was not enabled for w0. This resulted in the clock being
skewed so much that Amazon S3 would reject connections to it with a
`RequestTimeTooSkewed' error. If it wasn't for this, I would not have
noticed that w0's clock was drifting so much.

k0 does not seem to be affected (probably some change in upstream's
handling of ntp), but it doesn't hurt to add this configuration.