emacs: use `Input' instead of `Liberation Mono'
We use Input because for some reason, Liberation Mono's text height has
shrunk quite a bit because of a recent update (probably due to DPI
and/or font scaling issues). The font has become harder to read
(temporarily). So for now, use a customized version of the Input font,
available for free for private use from the Font Bureau.