nixos: disable pulseaudio
Pulseaudio was flaky (giving crackling sounds when using HTML5 video in
Firefox, although Chromium was fine with it). This had been the case for
probably 6 months. From Google searches it apparently happens only when
using a USB connection to an external DAC. Youtube recently went
HTML5-only, and that was the last straw. It might make sense to try
coming back to Pulseaudio next year to see if anything changed. Until
then, ALSA it is!