pulseaudio workaround for HTML5 video in Firefox
The trick is to set some settings in daemon.conf. The manpage at
pulse-daemon.conf(5) describes the default-fragment-size-msec and
default-fragments settings, which seem to affect our USB Schiit Bifrost
hardware. We can do a 1-liner change of just

    default-fragment-size-msec = 40

(default is 25), but this gives us a lot of latency. Using the bare
minimum fragments settings of

    default-fragments = 2
    default-fragment-size-msec = 1

does what we need. This seems to be a workaround against some internal
bug.

The big advantage of using pulseaudio over ALSA with the Schiit Bifrost
is that it gives us software volume controls for the CMUS player. This
is the primary reason for using pulse.