w0: intercept Print (F13) as Hyper_L, not Scroll_Lock (F14)
This is because F14 is mapped by OSX as "Decrease Brightness". If we
switch out of our VM and forget that we're not in it, then we end up
holding down F14, causing our Mac display brightness to dim. There is no
official Apple way to disable this behavior, so just use F13 instead.
Apparently, F13, F16-F19 are unmapped (F14/F15 are for controlling the
display brightness).