track uim files
Most of these are actually not that interesting, but certain files are
rather important; e.g., the `uim/customs/custom-ja-rk-rule.scm' file
contains hand-tuned conversion rules to convert sequences of ASCII
English letters to hiragana.

For the custom-ja-rk-rule.scm, by default it comes with many bindings.
We simply add on top c- and ch- bindings to allow for typing these
characters more easily on the ZQ layout, where the 'k' sits on the left
hand side on top of the vowels.