zsh: gdf: fix trailing newline chomping behavior
Instead of unconditionally deleting the final line all the time, only
delete totally empty lines.

This change was necessary because although git-diff (version 2.6.4,
Apple Git-63) always adds a blank line at the end, it does not do so
when the diff is that of fixing a

    "\ No newline at end of file"

instance by adding a newline. So, gdf would set $git_diff to be

    -xxx
    \ No newline at end of file

(because of the unconditional deletion of the last line) when really it
should look like

    -xxx
    \ No newline at end of file
    +xxx

.