zsh: fix mac machine aliases
In particular, using 'lmac' instead of linuss-air.ofc.... is more robust
because it uses the internal vboxnet0 connection instead of going over
the wire to the *.twinprime.com domain. We also don't need to specify
user "l" because the mac's username was changed from "larver" to "l"
a while ago.