emacs: help mode: use normal mode '<N>' instead of motion state '<M>'
Motion state is not the same as normal mode; in particular it does not
have any of the normal mode bindings that we already define in maps.el.
The problem is that help-mode by default runs in motion state, so that
keys such as TAB (which in normal mode run `other-window') do not have
the same effect. To get around this, we simply use normal mode for
help-mode; RET surprisingly works already in help-mode to follow
hyperlinks, so that's already probably 95% of help-mode's usefulness
anyway.