xmonad: add shortcut for emacs
The shortcut was removed in 6eede0b8e984415982905cec149e61ef2ea4de95, but it's
one I do really miss. The main reason I miss it is because oftentimes I switch
to a new blank workspace first, to put emacs there. Without this shortcut we
have to find a terminal first somewhere, invoke emacs, and then finally move
emacs to the blank workspace.