emacs: improve hydra bindings for helm
Previously we used the spacebar to bring up the hydra menu; this was not the
best configuration because it meant that for helm-ag, we could not do fuzzy
searches (helm-ag's "fuzzy" search works by treating search terms separated by
spaces as separate words, sort of like how in the terminal you can pipe each
word you want to search for separately into a new invocation of grep). We now
use the TAB key to do this.

The downside is that we can no longer bring up the action menu of helm directly;
still, this is not a huge loss because:

- we almost never use helm actions to begin with, and
- we can access the actions anyway with SPC once we are in hydra.

The only oddity is how we now have to map "TAB" to "<tab>" in the defhydra block
--- if we don't, the TAB key simply does not work in the hydra menu.