nixpkgs: disable profiling for all packages
Enabling profiling globally is not really needed any more as we move more
Haskell projects over to Stack (that is, we no longer use vanilla Nixpkgs
packages for Haskell development). Also, it cuts down the install time of
Haskell packages as we don't have to compile them from source (to get the
profiling builds of modules).