emacs: org: prettify syntax highlighting for source code blocks upon save
We use `after-save-hook' instead of `before-save-hook' because we want to keep
these automated changes (which amount to linting) atomically separate from
human-made changes --- esp. for source control management. To illustrate, if we
used `before-save-hook', we would have to sift through the linted changes vs the
human changes (if any) when using `git add -p'.