tig: make bindings resemble notmuch-emacs bindings
In our notmuch-emacs config, `(' and `)' move around emails, and `}' moves to
the next thread. The only thing that does not nicely map analagously is the `{'
key, which moves to the previous thread. This is because, while we can always
move "down" to the parent of a commit, it does not make sense to move to the
"child" of a commit, because a commit may have many children (all commits that
branched off from it).