offlineimap/notmuch: do not track tagging
Keeping the tagging information public feels a bit too personal (e.g., having
information about which emails to tag based on the name of the banks and other
financial institutions) --- so just keep the tagging part elsewhere. Other than
that, the only odd thing about this change is that we have to run the tag
decryption code in `secrets.py', which is a bit ugly, but I am too lazy to
create an entirely separate script just for this same purpose (to be called by
`presynchook', presumably).