offlineimap: expand scope of emails for deletion
Before we only deleted emails in the INBOX folder, but this meant that deleting
mailing list messages (filtered emails with the "inbox" tag removed) would never
get fully deleted. So, instead of a whitelist, we use a blacklist --- namely, we
only want to exclude sent messages from being deleted (as we want to save the
history of all sent messages).