emacs: when committing, move point to tho beginning of the buffer
Recall that we use emacsclient for writing commit messages. It appeared
before that sometimes when we commit, point would not appear at the
beginning of the buffer. This was because the COMMIT_EDITMSG buffer (as
a file) would get reused across multiple commits and, as Emacs thought
that it was the same buffer, it would restore point in the same location
as it was before (as if we were switching focus between windows back and
forth). The fix is to simply force Emacs to move point to the beginning
upon invoking the editor (in emacsclient-tty.sh).