xmonad: rename variable modm -> hypr
We use xmodmap outside of XMonad to create a Hyper key. This Hyper key
is set to the mod3Mask key by xmodmap, which we then pass on to XMonad
to designate as the modMask key (which we use pervasively in XMonad
everywhere as `modm'). Let's call it what it really is, instead of
hiding behind the `modm' moniker.

We also rename `mod1S' as `altS'.