xmonad: add 3rd dimension (Y-axis)
This as a major rewrite of xmonad.hs. Here are some of the highlights:

- added 3rd dimension (Y-axis)
  - can move across Y axis with H-M-{j,k} (all screens update!)
  - can shift windows across Y axis with H-M-S-{j,k} (focus follows
  - shifted window)
- standardize configuration on the 3-dimensional naming scheme (no more
  confusing VirtualWorkspaces and PhysicalWorkspaces and MyVWGroup)
- made H-S-n smarter by not shifting the focused window if that window
  - is the only window in the current workspace
- changed H-S-{h,l} so that after using the binding, focus stays with
  - the shifted window
- renamed all toplevel functions to have a "l_" prefix
- made some windows always spawn in particular ZGroup
  - this was previously a bit shoddy and felt broken
- added helper scripts for recompiling/restarting with `stack'
- remove obsolete Makefile (had not been used in years)

See the text toward the top for a more full-fledged discussion. The
TL;DR is that previously, each "group" was the Z axis, with the Xinerama
screens being the X axis; this patch adds a third (Y) axis. This change
was a result of me asking the question on Reddit [1], which itself was
inspired by a YouTube video [2].

[1]: https://www.reddit.com/r/xmonad/comments/7mawso/switch_workspaces_on_multiple_monitors_with_1/
[2]: https://youtu.be/w5_36BBGoU0?t=3m30s