xmonad: followup changes
This is a followup to 8b3c902d20d1f16543099b7739d31b425f0d1565. It has
not been fully tested yet, but it's better than the status quo because
the status quo is broken (at least for enif).

Among other things, we make NixOS-enabled hosts use the stack-compiled
version of xmonad. The enif changes are untested. We also favor using
~/bin instead of ~/.local/bin.