xmonad: make types stronger (do not use String where possible)
This way, switching around an X coordinate for a Y coordinate is
impossible, at least at type boundaries (e.g., when constructing XZY
from the (X, Z, Y) triplet).

It also results in a much cleaner design because we don't have to
manually construct/deconstruct a String type to get out the coordinate;
instead we rely on pattern matching for most tasks. The only exception
is when we have to read in a WorkspaceId.