xmonad: partially resurrect H-[0..6]/H-S-[0..6] bindings, for YCoords
Before we used the H-[0..9, F1-F12] (with and without Shift key)
bindings to view a particular Workspace or shift the current window to
that Workspace. In d38ab611040172980848cd100832b9e61f52f1d8 we deleted
these bindings because their use of W.greedyView and lack of
understanding of the Y-axis made them impractical.

This patch reintroduces the bindings, but with a couple changes. First,
the bindings view/shift to a YCoord. Second, we only use 7 numbers
instead of 22 (0..9 and F1-F12) because there are only 7 YCoords.

Lastly, before this patch, H-y (l_viewLastY) was limited to adjacent
YCoords because we had to cycle through each one in turn to get to our
target YCoord. This restriction is now lifted.