emacs: for tty, load elscreen explicitly
This has been an issue ever since we started using this script for doing
Git commits. The tab title that ELScreen gives us (which normally takes
up the very top line of the window) was not activated on startup, and
required us to move our cursor (to the beginning of the buffer) to
activate it.