xmonad: l_manageHook: use l_shiftAndViewAsHook instead of doShift
This new ManageHook not only shifts windows to their proper Workspace as
defined in l_manageHook, but also switches focus to it. This way, we can
invoke l_resetWindow on a focused window and maintain focus on it after
it is "reset" to a (possibly) different Workspace.

Because it is a ManageHook change, it affects not just the H-g
binding (l_resetWindow), but also all windows that are spawned at any
time (e.g., l_startupHook). So, when we log in, we will end up
"following" all windows that are spawned.