emacs: notmuch: simplify thread navigation
We don't use the S-<down>/S-<up> bindings any more. Also, preferring to
use J/K to do thread-level navigation and <down>/<up> to do
message-level navigation encourages us to just look at the heading of
each thread and to skip the entire thread if it's not that
interesting (J/K is easier to press than <down>/<up> --- on both QWERTY
and ZQ layouts).