xmonad: l_resetMouse: update mouse location for all non-mouse events
We add it to all keybindings in `l_keyBindings', so that all keybound
actions result in resetting the mouse. We also add it to `logHook',
because sometimes (when we create/delete windows *without* using our
keybindings) we want the mouse pointer to be reset. The idea to do use
the `logHook' is stolen from XMonad.Actions.UpdatePointer.