emacs: arjen-grey: use custom git-gutter colors
This patch is a bit unfortunate because arjen-grey theme is not the one
changing these colors. They are changed when we load in zenburn (that
theme sets these colors), so we have to overwrite them with the same
color information as in l-git-gutter.el.

It's probably not even zenburn's fault --- from zenburn's point of view
it is arjen-grey that lacks setting these colors.

The clean solution would be to have Emacs load themes but then
immediately undo the theme's code execution so that any newly-defined
faces get removed, so that loading a theme is a NOP as long as that
theme is not selected to be "active". When that happens, this patch will
not be necessary.