git_rebase_range: print hash of range's tip
Imagine we are on a branch called "dev" that is on top of master. We can
now do something like:

	$ git_rebase_range origin/master HEAD~3 HEAD

and it will print the new hash of the old HEAD after the rebase, so that
we can do

	$ git fetch . $hash:master
	$ git push origin master

to push up the last 4 changes on dev branch to master (where $hash is
what we print out to the user from git_rebase_range).

Also add a blurb for updating local/remote after we are done with the
rebase, if the revision anchor had a "/" in it.