emacs: fix package-archives
It turns out that `add-to-list' does not work for adding multiple
elements! We had not noticed this error sooner because it appears that
although the "melpa" archive did get added (adding hundreds of entries
to `list-packages'), the un-added "org" repository only adds 1 entry
"org" and so it was difficult to spot.