emacs: helm: set default size
This is a natural follow-up to a050d313d20b831a998d0dcc348b33ad2f306582.

NOTE: This makes it so that when the frame pops up in a non-XMonad
environment, the width and height are set accordingly. Under XMonad, the
frame (X window) is tiled, so these settings don't apply.