pulse: k0: set sample rate to 32-bits, 192kHz
The noticeable effect is that running `pacmd list-sinks` shows this:

	name: <alsa_output.usb-Schiit_Audio_Schiit_USB_Audio_Gen_2-00.iec958-stereo>
	driver: <module-alsa-card.c>
	flags: HARDWARE DECIBEL_VOLUME LATENCY FLAT_VOLUME DYNAMIC_LATENCY
	...
	sample spec: s32le 2ch 192000Hz

instead of

	name: <alsa_output.usb-Schiit_Audio_Schiit_USB_Audio_Gen_2-00.iec958-stereo>
	driver: <module-alsa-card.c>
	flags: HARDWARE DECIBEL_VOLUME LATENCY FLAT_VOLUME DYNAMIC_LATENCY
	...
	sample spec: s16le 2ch 44100Hz

so presumably it's better. I'm too lazy to test whether this actually
has any noticeable effect on sound quality, but it can't hurt.

The old settings around default-fragments and default-fragment-size-msec
no longer have any effect, at least on the 4.14.67 kernel, so they are
removed here as well.