xmonad: update dependencies
Use stack lts-13.4 (ghc 8.6.3) and HEAD.stack (1.9.3 x86_64 hpack-0.31.1), because for some mysterious reason the old stack binary started complaining with

	/nix/store/rjsymbdxlwmfbpasn0jik1w97wgfk3qj-glibc-2.27-bin/bin/ldconfig: Can't open cache file /nix/store/fivq0nbggp4y8mhy3ixprqd7qyn1hy2j-glibc-2.27/etc/ld.so.cache
	: No such file or directory
	I don't know how to install GHC on your system configuration, please install manually

To fix this, we updated the l-set-haskell packages in Nix to use
HEAD.stack instead of the old 1.7.1 version in the Nix nixos-18.09
channel. Then "stack setup", etc worked.