xmonad: do not wrap Y axis
In practice, it is not that useful to be able to switch to empty Y
coordinates, because most of the time we want to move around
already-populated Y coordinates. And so, we make the j/k bindings move
across already-populated Y coordinates only. Also, we disable wrapping
because cognitively it's much easier to just move to the very top or
bottom Y coordinate by holding down a j/k button, then to move again in
the opposite direction one keypress at a time (wrapping sort of disables
"tactile" feedback like this).

Also, use Return and BackSpace bindings for the old j/k behavior, and
not n/p (i.e., H-A-{n,p}), because they feel different against what we
already have for H-{n,p}.