xmonad: use custom (simpler) shell.nix
Instead of using

    cabal2nix --shell . > shell.nix

we instead use a custom shell.nix file that pulls in the Nix packages at
a particular channel. Although the channel name "19.03" will keep
moving, it is fine because it has been tested upstream. We could use a
SHA as an argument to fetchGit, but it is unnecessary.

This has the advantage of letting stack download all of the latest
Haskell packages (instead of relying on the pre-built Nix ones). This
makes more sense because if we're using the stack LTS resolver, we might
as well try to defer as much to stack's package database as possible
(instead of using Nixified versions as per cabal2nix).

Also update LTS to 13.29 (GHC 8.6.5).

The tip to use a custom shell.nix is from
https://stackoverflow.com/a/56180220/437583.