xmonad: add libraries explicitly
This is a result of us dropping "HEAD.stack" from l-set-haskell in our
nixpkgs. That version of stack comes with the gmp and libffi packages,
and so removing HEAD.stack from l-set-haskell meant dorpping these
packages as well. This means we have to add them explicitly here. The
reason why we didn't have to do this before was probably because we use
an impure nix environment in stack.yaml.