emacs: kakakpo: enable "global-git-commit-mode" for git commits
The trouble here is that `git-commit` is its own package that Magit and
other packages use, and is not a major mode. Instead of autoloading this
package, we invoke it here from kakapo mode's settings. This is because
the kakapo settings are a centralized area that deals with all types of
specialized file paths, etc.

git-commit mode uses the `with-editor` ("WE" in the lighter) package.

Anyway, this change should make it so that the emacsclient always picks
up the mode. It's been something of a hit-or-miss.