emacs: delete minibuffer frame on exit
This removes a longstanding bug where sometimes opening files with the
emacsclient (to be opened as a graphical X window, not in the terminal)
via our univ_open zsh function would not work, where the buffer would
instead be opened in an existing frame. The existing frame in this case
happened to be the *invisible* (unraised) minibuffer frame. This
resulted in essentially a NOP.

It turns out that this behavior was triggered whenever there happened to
be a minibuffer frame that was opened from the initial emacsclient
graphical window.