emacs: l/quit-buffer: close other buffers for graphical clients
The if-guard for (display-graphic-p) is to make sure that we try to
close other buffers loaded in the daemon only when we close a graphical
client. This is to cover the case where we are editing a commit message
from within a terminal emacsclient session. That is, if we are exiting a
terminal emacsclient as part of saving a Git commit message, we would
not want to disturb any open graphical emacsclients that may be open.

Likewise, if we do delete buffers, we only delete those ones that are
not used for Git.