k0: xorg: set blank time to 1hr instead of 10 minutes
By default, if we do `xset -q`, we get

    ...
    Screen Saver:
    prefer blanking:  yes    allow exposures:  yes
        timeout:  600    cycle:  600
    ...

The screen blanking (in seconds) currently makes the monitors enter a
sleep after 10 minutes of inactivity. This is too short, so go for 1
hour. The main thing to note is that the X configs are in minutes, not
seconds.

To set this manually, do `xset s 3600 3600`.