xmonad: H-{h,l}, H-S-{h,l}: auto-promote hidden workspace
Make it so that both the old and new X coordinate is not left in a state
where an empty workspaces is the one being displayed (no windows), when
there is an idle nonempty one hidden underneath it in the Z coordinate.

This makes it so that we don't have to press xK_n manually to get to the
hidden nonempty workspace.