emacs: don't cycle through fonts if there is only 1 font
For whatever reason, if we cycle through this list at all, the choice
for "Input Mono Condensed" always results in picking plain "Input Mono"
or "Input Mono Narrow" (it's much wider when we call this function on
Mac). Since Mac only uses 1 font at all, we make this function do NOP if
there are no fonts to cycle through.