emacs: disable backups on Mac for better org-mode interop
For some reason, when org-mode links to other *.org files become
automatically updated to be a link to the backup of that other file,
whenever that other file is edited within the same Emacs session. This
is really annoying as it breaks the links, and also sometimes results in
the original file being completely deleted on disk while leaving the
backup OK.

Disabling backups resolves the issue completely, and it's an acceptable
workaround given how frequently I create Git commits and how rarely I
ever use Emacs' backups of files. It might be worth revisiting in the
future, but I'm not in a rush.