nixos: k0: use xrandr instead of xinerama
This allows us to use things like alacritty, which requires the xrandr
extension to be enabled (turning on xinerama disables xrandr).

For xmonad, we have to spawn things slightly differently because the
screens are ordered slightly differently now (for whatever reason).

As a pleasant surprise, YouTube in Chromium now correctly only
full-screens videos to stay within a single screen (instead of trying to
fit across all 4 monitors).