emacs: deprecate elscreen in favor of native tabs
Emacs 27.1 has support for tabs natively. Use this instead of the
"elscreen" third-party package.

The functionality is virtually identical. The one difference is that the
tab bar is no longer embedded inside the buffer, and is actually a
separate GUI element. This means that splitting the same buffer with
multiple windows no longer creates multiple tab bars (a
longstanding (minor) annoyance).