zsh: emacs: ignore snippets file
We use the snippets.org file to record old entries; these should not show up on
the agenda.

NOTE: We prefer not to use the default org-archive-subtree-default command
because it is a bit unwieldy. In the future, we can try modifying the archiving
function as in [1] to change how headings get archived (perhaps to different
files based on dates).

[1]: https://emacs.stackexchange.com/questions/24992/archiving-completed-tasks-to-an-archive-file-based-on-the-current-date-in-org-mo