zsh: __l_prompt_tick: resume tick after pressing ENTER
If we run ENTER repeatedly, accept-line is the latest $WIDGET to be actiivated,
and the prompt no longer ticks. To guard against this overbroad behavior, we
just additionally require that the BUFFER also be populated in order for the
tick to block. This combination more truly captures the state of multi-line
input, which is what we want to catch here.