alacritty/zsh: move C-h, C-l bindings to ZSH
The old configuration had the disadvantage (as noted in the comments):

    NOTE: These keybindings for C-h and C-l interfere with terminal emacs'
    next/prev tab. However we don't use terminal emacs enough to justify using a
    different set of bindings.

However this was incorrect, because setting it at the alacritty (base) layer
meant that it would get interpreted in a ZSH-like way for all instances,
including inside of terminal emacs and wherever else. Delegating these bindings
to ZSH directly allows alacritty to properly send C-h and C-l as-is when we are,
e.g., SSH-ed into a remote machine.