alacritty: mac: enable all M-[a-z], M-S-[a-z] bindings
On Mac, the default US keyboard binds the M-[a-z] and M-S-[a-z] combinations for
rarely-used Unicode symbols. If we use Alacritty to SSH into a tmux session on
our k0 linux box, and from there use an emacsclient session, we end up sending
over these Unicode symbols with the aforementioned key combos.

Setting the combinations explicitly here disables the Unicode silliness.