emacs: make git-gutter refresh more frequently
Terminal emacs doesn't get the frame focus hook events, so we have to use a
different hook. We just advise "select-window", because it is used every time we
choose a window.

This is slightly better than advising "other-window" which we have mapped to the
TAB key in normal mode, because it covers the case where we switch tabs with C-h
and C-l, which are not covered by "other-window". (This is just my theory, but
it makes sense to me and has been observed in practice.)