emacsclient: prefer find-file
If we use find-file-other-tab, we end up with 2 tabs --- one for
"*scratch*" and the other for the actual file we want to edit. Because
we'll really only have 1 file we care about whenever we invoke this,
there is no need to open that file in a second tab.

For the record, the idea around using find-file-other-tab came from the
scenario where we wanted to reuse an existing frame, because in that
scenario, we didn't want to change buffers on any existing tab's
windows.