alacritty: set static terminal cursor color
The default is to make the cursor change colors based on the current
cell's foreground/background. However this runs into the issue in
terminal Emacs org-mode with hidden leading stars (either with
org-startup-indented or org-hide-leading-stars); in this situation,
if the cursor is over a hidden leading asterisk, because the foreground
color matches the background color (to hide the asterisk), Alacritty
also hides the cursor completely. The solution is to just always force
Alacritty to display the cursor in a particular color. We choose white
with black text here.