emacs: more gracefully handle quitting buffers
The issue with the previous setting was that we did not actually kill
the COMMIT_EDITMSG and git-rebase-todo buffers after exiting them with
l/quit-buffer. So instead make sure to invoke either with-editor-cancel
or with-editor-finish for these buffers when quitting from them so that
the buffers are removed from buffer-list when we're done.