emacs: fix wrong usage of mapcan
The problem is that the surrounding "or" function expects a list
sequence. "mapcar" gives us this sequence, but using "mapcan" tried to
un-list-ize each thing inside the list; this was problematic because
each thing inside the list was already not a list. So we got this error:
"wrong type argument: consp, 0" because mapcan tried to convert the bare
"0" into a non-list.

We should have always just used "mapcar" to begin with. It is a mystery
how this bug survived for so long undetected. FWIW this code path was
only triggered by using a graphical Emacs, and I haven't used a
GUI-based Emacs in a while, so that may be why.